Revert "Solving #144 and #145"
