+++ TalkIn/app/build.gradle
@@ -41,9 +41,7 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'com.google.firebase:firebase-auth-ktx:23.0.0'
     implementation 'com.google.firebase:firebase-firestore-ktx:23.0.3'
-    implementation "androidx.viewpager2:viewpager2:1.0.0"
-    implementation 'com.google.android.material:material:1.3.0'
-    implementation 'com.github.bumptech.glide:glide:4.12.0'
+
     implementation 'com.google.firebase:firebase-database:21.0.0'
     implementation 'androidx.gridlayout:gridlayout:1.0.0'
     implementation 'androidx.activity:activity:1.8.0'
