--- TalkIn/app/src/main/java/com/example/talk_in/CameraFragment.kt
@@ -1,92 +1,0 @@
-import android.Manifest
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.graphics.Bitmap
-import android.os.Bundle
-import android.provider.MediaStore
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.RecyclerView
-import com.example.talk_in.R
-
-class CameraFragment : Fragment() {
-
-    private lateinit var recyclerView: RecyclerView
-    private lateinit var loadingImage: ImageView
-    private lateinit var capturedImageView: ImageView
-    private lateinit var textView: TextView
-
-    companion object {
-        private const val REQUEST_CAMERA_PERMISSION = 100
-        private const val REQUEST_CAMERA = 101
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        val view = inflater.inflate(R.layout.fragment_camera, container, false)
-
-        recyclerView = view.findViewById(R.id.gallery_recycler_view)
-        loadingImage = view.findViewById(R.id.loading_image)
-        capturedImageView = view.findViewById(R.id.captured_image)
-        textView = view.findViewById(R.id.send_image_button)
-
-        checkCameraPermission()
-
-        return view
-    }
-
-    private fun checkCameraPermission() {
-        if (ContextCompat.checkSelfPermission(
-                requireContext(), Manifest.permission.CAMERA
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            ActivityCompat.requestPermissions(
-                requireActivity(),
-                arrayOf(Manifest.permission.CAMERA),
-                REQUEST_CAMERA_PERMISSION
-            )
-        } else {
-            openCamera()
-        }
-    }
-
-    private fun openCamera() {
-        val cameraIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
-        if (cameraIntent.resolveActivity(requireActivity().packageManager) != null) {
-            startActivityForResult(cameraIntent, REQUEST_CAMERA)
-        }
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == REQUEST_CAMERA && resultCode == AppCompatActivity.RESULT_OK) {
-            val imageBitmap = data?.extras?.get("data") as? Bitmap
-            capturedImageView.setImageBitmap(imageBitmap)
-            capturedImageView.visibility = View.VISIBLE
-            textView.visibility = View.VISIBLE
-            loadingImage.visibility = View.GONE
-        }
-    }
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int, permissions: Array<out String>, grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (requestCode == REQUEST_CAMERA_PERMISSION) {
-            if ((grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED)) {
-                openCamera()
-            } else {
-                // Permission denied
-            }
-        }
-    }
-}
