--- TalkIn/app/src/main/res/layout/fragment_camera.xml
@@ -1,43 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#F0EEEF">
-
-    <ImageView
-        android:id="@+id/captured_image"
-        android:layout_width="match_parent"
-        android:layout_height="400dp"
-        android:layout_alignParentTop="true"
-        android:scaleType="centerCrop"
-        android:visibility="gone" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/gallery_recycler_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/captured_image"
-        android:layout_marginTop="8dp"
-        android:visibility="gone" />
-
-    <TextView
-        android:id="@+id/send_image_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/gallery_recycler_view"
-        android:layout_marginTop="30dp"
-        android:clickable="true"
-        android:gravity="center"
-        android:text="Send Image"
-        android:textColor="#1E1F22"
-        android:textSize="16sp"
-        android:textStyle="bold"
-        android:visibility="gone" />
-
-    <ImageView
-        android:id="@+id/loading_image"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/camer"
-        android:layout_marginBottom="200dp"
-        android:visibility="visible" />
-</RelativeLayout>
